requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alembic==1.13.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
attrs==25.4.0
bitarray==3.8.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cytoolz==1.1.0
ecdsa==0.19.1
eth-account==0.10.0
eth-hash==0.7.1
eth-keyfile==0.9.1
eth-keys==0.7.0
eth-rlp==1.0.1
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==6.0.0b1
facenet-pytorch==2.5.3
fastapi==0.110.0
filelock==3.20.0
frozenlist==1.8.0
fsspec==2025.12.0
h11==0.16.0
hexbytes==0.3.1
httptools==0.7.1
idna==3.11
Jinja2==3.1.6
joblib==1.3.2
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lru-dict==1.2.0
Mako==1.3.10
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.7.1
networkx==3.6.1
numpy==1.26.4
opencv-python==4.9.0.80
pandas==2.2.1
parsimonious==0.10.0
pillow==10.2.0
propcache==0.4.1
protobuf==6.33.5
psycopg2-binary==2.9.9
py-ecc==8.0.0
pyasn1==0.6.2
pycparser==3.0
pycryptodome==3.23.0
pydantic==2.6.4
pydantic-settings==2.2.1
pydantic_core==2.16.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2025.2
pyunormalize==17.0.0
PyYAML==6.0.3
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
rlp==4.1.0
rpds-py==0.30.0
rsa==4.9.1
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.17.0
SQLAlchemy==2.0.29
starlette==0.36.3
sympy==1.14.0
threadpoolctl==3.6.0
toolz==1.1.0
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.29.0
uvloop==0.22.1
watchfiles==1.1.1
web3==6.15.1
websockets==16.0
yarl==1.22.0
